careconnect/hms/hospital/requirements.txt
absl-py==2.1.0
altair==5.2.0
annotated-types==0.6.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
babel==2.16.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
capstone==5.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colored-traceback==0.3.0
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.6
django_db==0.0.7
dnspython==2.6.1
executing==2.0.1
fastapi==0.115.0
fastjsonschema==2.20.0
filelock==3.13.1
Flask==3.0.2
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
flatbuffers==24.3.25
fonttools==4.47.2
fqdn==1.5.1
fsspec==2024.2.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.7
importlib-metadata==7.0.1
intervaltree==3.1.0
ipykernel==6.29.3
ipython==8.22.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
mysql-connector-python==9.1.0
mysqlclient==2.2.6
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.3
opt-einsum==3.3.0
optree==0.12.1
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
plumbum==1.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.7
pure-eval==0.2.2
pwntools==4.11.1
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycparser==2.21
pycryptodome==3.19.1
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
pyelftools==0.30
Pygments==2.17.2
pymongo==4.6.3
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
ROPGadget==7.4
rpds-py==0.18.0
rpyc==5.3.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
Send2Trash==1.8.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.38.6
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.3.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
types-python-dateutil==2.9.0.20240906
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
unicorn==2.0.1.post1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.6
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
widgetsnbextension==4.0.13
wrapt==1.16.0
xgboost==2.0.3
zipp==3.17.0
